refactor(editor): wire NoteEditorScreen to extracted hooks
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
